b1.21.0 - refactor shared.go
