+++ src/utils.go
@@ -1,0 +1,215 @@
+package main
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"log"
+	"strings"
+	"time"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/microsoftgraph/msgraph-sdk-go/models/odataerrors"
+)
+
+// isRetryableError determines if an error is transient and worth retrying.
+// Returns true for network timeouts, Graph API throttling (429), and service
+// unavailability (503) errors.
+func isRetryableError(err error) bool {
+	if err == nil {
+		return false
+	}
+
+	// Check for context cancellation - never retry these
+	if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
+		return false
+	}
+
+	// Check for Azure SDK response errors
+	var respErr *azcore.ResponseError
+	if errors.As(err, &respErr) {
+		if respErr.StatusCode == 429 || respErr.StatusCode == 503 || respErr.StatusCode == 504 {
+			return true
+		}
+	}
+
+	// Check error message for common transient patterns
+	errMsg := strings.ToLower(err.Error())
+	transientPatterns := []string{
+		"timeout",
+		"connection reset",
+		"connection refused",
+		"temporary failure",
+		"try again",
+		"i/o timeout",
+		"no such host",
+		"network is unreachable",
+	}
+
+	for _, pattern := range transientPatterns {
+		if strings.Contains(errMsg, pattern) {
+			return true
+		}
+	}
+
+	return false
+}
+
+// enrichGraphAPIError enriches Graph API errors with additional context,
+// particularly for rate limiting scenarios. It detects rate limit errors (429)
+// and extracts the Retry-After header if available.
+func enrichGraphAPIError(err error, logger *CSVLogger, operation string) error {
+	if err == nil {
+		return nil
+	}
+
+	// Check if this is an OData error from Microsoft Graph
+	var odataErr *odataerrors.ODataError
+	if !errors.As(err, &odataErr) {
+		// Not an OData error, return as-is
+		return err
+	}
+
+	// Extract error details if available
+	if odataErr.GetErrorEscaped() == nil {
+		return err
+	}
+
+	errorInfo := odataErr.GetErrorEscaped()
+	code := ""
+	message := ""
+
+	if errorInfo.GetCode() != nil {
+		code = *errorInfo.GetCode()
+	}
+	if errorInfo.GetMessage() != nil {
+		message = *errorInfo.GetMessage()
+	}
+
+	// Handle rate limiting (429 TooManyRequests)
+	if code == "TooManyRequests" || code == "activityLimitReached" {
+		log.Printf("[WARN] Graph API rate limit exceeded during %s (code: %s)", operation, code)
+
+		// Try to extract Retry-After header
+		retryAfter := ""
+		if odataErr.GetResponseHeaders() != nil {
+			if retryHeaders := odataErr.GetResponseHeaders().Get("Retry-After"); len(retryHeaders) > 0 {
+				retryAfter = retryHeaders[0] // Get first value
+				log.Printf("[INFO] Rate limit retry guidance available: retry after %s seconds", retryAfter)
+			}
+		}
+
+		// Build enriched error message
+		enrichedMsg := fmt.Sprintf("rate limit exceeded during %s", operation)
+		if retryAfter != "" {
+			enrichedMsg += fmt.Sprintf(" (retry after %s seconds)", retryAfter)
+		}
+		enrichedMsg += ". Consider: 1) Reducing request frequency, 2) Implementing exponential backoff, 3) Reviewing API throttling limits"
+
+		return fmt.Errorf("%s: %w", enrichedMsg, err)
+	}
+
+	// Handle other service errors (503, 504)
+	if code == "ServiceUnavailable" || code == "GatewayTimeout" {
+		log.Printf("[WARN] Graph API service error during %s (code: %s, message: %s)", operation, code, message)
+		return fmt.Errorf("service temporarily unavailable during %s (code: %s): %w", operation, code, err)
+	}
+
+	// For other OData errors, log details for debugging
+	if code != "" {
+		log.Printf("[DEBUG] Graph API error during %s (code: %s, message: %s)", operation, code, message)
+	}
+
+	return err
+}
+
+// retryWithBackoff wraps an operation with exponential backoff retry logic.
+func retryWithBackoff(ctx context.Context, maxRetries int, baseDelay time.Duration, operation func() error) error {
+	var lastErr error
+
+	for attempt := 0; attempt <= maxRetries; attempt++ {
+		// Execute the operation
+		lastErr = operation()
+
+		// Success - return immediately
+		if lastErr == nil {
+			if attempt > 0 {
+				log.Printf("Operation succeeded after %d retries", attempt)
+			}
+			return nil
+		}
+
+		// Check if error is retryable
+		if !isRetryableError(lastErr) {
+			// Non-retryable error - fail immediately
+			return lastErr
+		}
+
+		// Last attempt failed - return error
+		if attempt == maxRetries {
+			return fmt.Errorf("operation failed after %d retries: %w", maxRetries, lastErr)
+		}
+
+		// Calculate exponential backoff delay (cap at 30 seconds)
+		delay := baseDelay * time.Duration(1<<uint(attempt))
+		if delay > 30*time.Second {
+			delay = 30 * time.Second
+		}
+
+		log.Printf("Retryable error encountered (attempt %d/%d): %v. Retrying in %v...",
+			attempt+1, maxRetries, lastErr, delay)
+
+		// Wait with context cancellation support
+		select {
+		case <-ctx.Done():
+			return fmt.Errorf("retry cancelled: %w", ctx.Err())
+		case <-time.After(delay):
+			// Continue to next retry attempt
+		}
+	}
+
+	return lastErr
+}
+
+// maskSecret masks a secret for display
+func maskSecret(secret string) string {
+	if len(secret) <= 8 {
+		return "********"
+	}
+	// Show first 4 and last 4 characters
+	return secret[:4] + "********" + secret[len(secret)-4:]
+}
+
+// maskGUID masks a GUID showing only first and last 4 characters
+func maskGUID(guid string) string {
+	if len(guid) <= 8 {
+		return "****"
+	}
+	return guid[:4] + "****-****-****-****" + guid[len(guid)-4:]
+}
+
+// Helper: Return default string if empty
+func ifEmpty(s, defaultVal string) string {
+	if s == "" {
+		return defaultVal
+	}
+	return s
+}
+
+// Helper: Truncate string with ellipsis
+func truncate(s string, maxLen int) string {
+	if len(s) <= maxLen {
+		return s
+	}
+	return s[:maxLen] + "..."
+}
+
+// Helper function to create int32 pointer
+func Int32Ptr(i int32) *int32 {
+	return &i
+}
+
+// pointerTo is a generic helper function to create pointers to values
+func pointerTo[T any](v T) *T {
+	return &v
+}
